src/pyedb/workflows/job_manager/requirements.txt
aiohttp>=3.8.0
asyncio>=3.4.3
colorama>=0.4.6; platform_system=="Windows"
numpy>=1.24.0
pandas>=1.5.0
psutil>=5.9.0
pyedb>=0.6.0
python-socketio[client]>=5.8.0
# Windows-specific requirements
pywin32>=306; platform_system=="Windows"
streamlit>=1.28.0
